namespace vectorsearch
