+++ lib/langchain/vectorsearch/qdrant.rb
