+++ spec/langchain/vectorsearch/pgvector_spec.rb
